// src/4_entities/Photo/model/types/ArticlePhoto.ts
import { ArticleCategoryType } from '4_entities/Category';
import { DetailMainType } from '5_shared/types/DetailMainContent';

interface ArticlePhotoType {
    id: number;
    slug: string;
    title: string;
    locale: string;
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    main: DetailMainType;
    tags?: {
        data: ArticleCategoryType[];
    };
    category?: {
        data: ArticleCategoryType;
    };
}

export type {
    ArticlePhotoType,
};